Midi scummvm refactor (#167)
* [wip] mididrv files split/refactor

* [wip] MidiDriver_ADLIB split/refactor

* gcc build fix

* code rev

* code rev
